Improve how we listen to branch updates, again.
You can now specify a list of branches to listen to, in an array or
you may also specify a branch to ignore by prefixing "!" or "-".

Now to introduce a more advanced *matches than previously allowed.
Before you were limited to * (everything). Now you may use *(prefix)
and/or (suffix)* for *endsWith, startsWith* or *contains* matching.